extract plugins from catalog
