+++ catalog/internal/pluginmanager/plugins_test.go
@@ -1,4 +1,4 @@
-package plugins
+package pluginmanager
 
 import (
 	"os"
