--- catalog/internal/plugins/config.go
@@ -1,15 +1,0 @@
-package plugins
-
-import (
-	"strings"
-)
-
-type Config struct {
-	PluginsDir string
-}
-
-func LoadConfig(pluginsDir string) Config {
-	return Config{
-		PluginsDir: strings.TrimSpace(pluginsDir),
-	}
-}
